move module to common.salers
move to common.scalers
